tiac_magpie: no entropy generator as default
It is not necessary to enable the on-chip entropy generator
as board default as long as only a few applications will need
it and also one of the ARM architecture test cases will fail
with already enabled hardware entropy generator.

The failed ARM architecture test case is:

  zephyr/tests/arch/arm/arm_irq_vector_table

Each sub-system or application that depends on the on-chip
entropy generator have to enable it for it self. For example
thus is the case when the network would be activated:

  zephyr/subsys/net/Kconfig:

    config NETWORKING
           select ENTROPY_GENERATOR

  zephyr/soc/arm/st_stm32/stm32f7/Kconfig.defconfig.stm32f767xx:

    if ENTROPY_GENERATOR
    config ENTROPY_STM32_RNG
           default y
    endif # ENTROPY_GENERATOR

issues: #3
Signed-off-by: Stephan Linz <linz@li-pro.net>