soc: arm: add support for STM32F777
As long as bridle is not able to extend the original
SOC Kconfig environment in Zephyr the support for
STM32F777 is more an variation of the STM32F767. As an
result both SOC configuration have to enabled on same
time: CONFIG_SOC_STM32F767 and CONFIG_SOC_STM32F777.

issues: #4
Signed-off-by: Stephan Linz <linz@li-pro.net>