doc: cmake: add dodgy mode
Some docsets may not be able to get rid of doc warnings, specially those
that are external and not controlled by Nordic. This mode allows makes
SPHINXOPTS_EXTRA to not be passed to Sphinx, used by the CI to enable -W
mode.

Copy-of: https://github.com/nrfconnect/sdk-nrf/commit/01cd1cd
Signed-off-by: Gerard Marull-Paretas <gerard.marull@nordicsemi.no>
Signed-off-by: Stephan Linz <linz@li-pro.net>