doc: link-roles: clean up setup() method
The setup() method is trying to find the Bridle repository by name,
but it's not passing allow_paths=False to the west method it uses to
search.

That may lead to unpredictable results depending on what the current
working directory is in the calling environment.

For robustness, disallow paths and make sure we are only searching for
the Bridle project by name.

Add some more comments to explain what is going on and clean up the
empty string handling while we're here.

issues: #49

Copy-of: https://github.com/zephyrproject-rtos/zephyr/commit/8eeac3e
Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>
Signed-off-by: Stephan Linz <linz@li-pro.net>