modules: Remove unneeded Kconfig from modules folder
This folder is intended for glue code for 3rd party
zephyr modules where this glue code is not wanted inside
the upstream code base.
Integration into Kconfig happens automatically though through
an automated discovery process and manually pulling this file in
is not needed and may lead to undefined behaviour during the kconfig
stage.