doc: scripts: merge Kconfig search database
Kconfig database was not part of the merged search index due to its
massive size. However, this is no longer the case with the new Kconfig
search extension. With this change the global search will also work when
browsing the kconfig docset.

issues: #59

Copy-of: https://github.com/nrfconnect/sdk-nrf/commit/04ea8ac
Signed-off-by: Gerard Marull-Paretas <gerard.marull@nordicsemi.no>
Signed-off-by: Stephan Linz <linz@li-pro.net>