doc: add "go back to" page in all API documentations
The HTML pages generated by Doxygen are now based on a corresponding
and customized Doxygen layout XML file: "standalone" for an detached
Doxygen run and "zephyr-doxyrunner" for the indirect Doxygen run
through the Sphinx extension "zephyr.doxyrunner". The respective
layout presents a new site with links back to the respective other
documentation sets.

Signed-off-by: Stephan Linz <linz@li-pro.net>