tests: clone and adapt I2C slave test
Since we had been patched the original Zephyr I2C slave
test suite again and again at upstream Zephyr source code
level. To follow those is not a sustainable approach and
we want to avoid it with this cloned fork (still using
the original Zephyr source code), even if it seems a bit
unusual and strange.

issues: #64

Signed-off-by: Stephan Linz <linz@li-pro.net>