doc: add standalone Bridle API documentation
The Bridle API documentation is now not longer an out of
Sphinx tree build step. Now it uses the 'zephyr.doxyrunner'
extension for Sphinx in a proper and intended way.

Beside this and for rapid prototyping the new CMake target
'bridle-doxygen' is abel to render a standalone Doxygen API
documentation.

- bumb Doxygen minimum version to 1.9.1, as the used CSS
  theme 'doxygen-awesome-css' depends on Doxygen greater
  than 1.9.1 -- tested with Doxygen until 1.9.5
- copy in all 'doxygen-awesome-css' 1.6.1 theme artifacts
  for a 'sidebar-only' presentation with darkmode enabled
- extend the 'doxygen-awesome-css' theme with the specific
  sidebar background color for Bridle (docset distinction)

issues: #64

Signed-off-by: Stephan Linz <linz@li-pro.net>