shield: add Grove Wiring Test Bed shield
This shield is less a plug-on module in the conventional sense than
more a wiring for interconnection of certain signals. It requires a
board, maybe with additional connected shields, which provides a
configuration that allows:

- one of Grove LED Shields for digital data output line
- one of Grove Button Shields for digital data input line

When the board is not equipped with the required shield connectors,
but instead an Arduino R3 header, then the Grove Base Shield V2 can
be plugged in between.

Signed-off-by: Stephan Linz <linz@li-pro.net>