ci: github: add integration tests for shields
Now we have two stages in parallel, the "qa-shield-integration" stage
that runs on a common Linux system on GitHub with the help of Zephyr's
Docker container image "ghcr.io/zephyrproject-rtos/ci" v0.24.13 for
builds with the Tephyr SDK v0.15.2 and the "qa-target-integration",
former "qa-integration", that runs on our dedicated in-house machine
with direct access to test hardware.

Also the direct call to the Twister script was replaced by the new
"west twister" entrypoint. That allows to skip the explicit setup of
the two environment variables ZEPHYR_BASE and BRIDLE_BASE to prepare
the Twister test run sessions.

Signed-off-by: Stephan Linz <linz@li-pro.net>