boards: add downstream copy of Seeeduino XIAO
This copies the upstream board configuration from Zephyr into
the downstream environment of Bridle for further changes and
extensions. This is a preperation for development on this board
without dependencies to upstrem Zephyr.

All results from the Bridle downstream development should be
contribute to upstream Zephyr as soon as possible.

Signed-off-by: Stephan Linz <linz@li-pro.net>