west: remove 'civetweb' from import allowlist
Zephyr v3.2 removed this obsolete code. The CivetWeb [1] code has
gone unmaintained and bugs continue to be reported against it on
Zephyr. We do not have the resources as a project to maintain this
in "odd fixes" mode, and nobody has stepped up to maintain it [2],
so sadly this must be removed for now, see:

https://github.com/zephyrproject-rtos/zephyr/pull/46746

[1] https://github.com/civetweb/civetweb
[2] https://lists.zephyrproject.org/g/devel/message/8466

issues: #105
Signed-off-by: Stephan Linz <linz@li-pro.net>