shields: revert CAN timing adjustments for tests
This reverts commit ad49608981d9d7e299851e07652084c71d9b4ece.

Since the beginning, the adjustments of the CAN timing for board
TiaC Magpie STM32F777NIHx as a shield was more a workaround than
a clean and correct abstraction of this problem. With introduction
of snippets as part of the Zephyr build process, this way over a
shield can now be abandoned and the general-purpose snippet
'can-timing-adj' have to be used instead.

Signed-off-by: Stephan Linz <linz@li-pro.net>