feat: upgrade CMake
To interactively debug CMake version 3.27 is required. As
this container is based on the Zephyr SDK which is based
on an Ubuntu LTS version, CMake is too old to support
debugging.

Therefore we install CMake from their repository and add
the CMake extension to the devcontainer file for auto-install
it.

Signed-off-by: Andreas Kurz <andreas.kurz@methodpark.de>