tests: x_grove_testbed: use integration platform tests
Together with the new global quarantine list all specific
test cases are now more generic and independent of special
white or black lists. When ever an exclusion will be needed
this rule have to add to the new global quarantine list.

In addition, the "Grove Base V2" shield constellation
is also being tested on the Seeed Studio Lotus board.

Signed-off-by: Stephan Linz <linz@li-pro.net>