[WIP] Add more robot test files in documentation
We use the repo https://github.com/RIOT-OS/RobotFW-tests to
add more robot files to the documentation.

The problem is that some of the robot files have the same
description. During the build of the documentation we get
some warnings with 'duplicate label' and the link are not
working correctly.

Signed-off-by: Kerstin Jenett <kerstin.jenett@navimatix.de>