cmake: toolchain: early exit for non Zephyr variants
In the case of Zephyr toolchain variants other than 'zephyr', Bridle's
Zephyr SDK version guard should return as early as possible and not
qualify the specified Zephyr SDK version range. For example, if the
variant 'ZEPHYR_TOOLCHAIN_VARIANT' is set to 'gnuarmemb', Bridle will
not be able to qualify the correct version, which must be specified
by 'GNUARMEMB_TOOLCHAIN_PATH'.

This job must be monitored and protected by the Zephy CMake extensions.

Signed-off-by: Stephan Linz <linz@li-pro.net>