ci: labeler: fix breaking changes in V5
The ability to apply labels based on the names of base and/or head
branches was added. The match object for changed files was expanded
with new combinations in order to make it more intuitive and flexible.
As a result, the configuration file structure was significantly
redesigned and is not compatible with the structure of the previous
version.

Signed-off-by: Stephan Linz <linz@li-pro.net>