samples: ubx_gnss: Use logging service
This converts the uGnssPosGetStreamed callback to use the Zephyr
logging OS service, thus we get the uptime also printed out on
console and we are able on runtime to diable or maybe deligate
the callback output to other stream channels then the console.

Also error messages converted to use the shell_error() wrapper.

Signed-off-by: Stephan Linz <linz@li-pro.net>