drivers: add SC16IS75x series MFD core driver
This driver manages the MFD pseudo-bus on which child devices can
be placed, and which itself is placed on either an SPI or I2C bus.

The number of and ids of channels on a given bridge are determined
at build time based on devicetree information.

Most bus communication is performed using blocking calls, but to
perform interrupt handling some non-blocking bus operations are
needed. If the underlying bus supports asynchronous/callback based
APIs, these are utilized. Otherwise, each driver instance creates
a separate work queue for offloading bus transfers without blocking
the system work queue.

Signed-off-by: Sarah Renkhoff <sarah.renkhoff@navimatix.de>
Signed-off-by: Stephan Linz <linz@li-pro.net>