ci: github: qa-integration: restructuring the examples
In addition to Bridle's own examples, it should also be possible
to build specific Zephyr upstream examples with an alternative
configuration base with the Bridle-owned specifications. For this
purpose, the new tags `zephyr` and `bridle` must be introduced and
respected. For the GitHub workflows, this makes the explicit naming
of individual test roots unnecessary. The selection is made
exclusively via the newly introduced tags.

Signed-off-by: Stephan Linz <linz@li-pro.net>