samples: helloshell: avoid RAM overflow on qemu_cortex_m0
Since Zephyr upstream PR 74435, the sensor shell commands rely on
a complete integration of the sensor ASYNC API incl. RTIO subsystem
with a thread pool for a dedicated RTIO work queue. This exceeds
the 16KB RAM of the Cortex M0 emulator by far. The only option
available at the moment is to disable the sensor shell commands.

Signed-off-by: Stephan Linz <linz@li-pro.net>