drivers: i2c_sc18is604: refactoring I2C controller
On interrupt callback registration the complete 32-bit mask should
be used.

The result code is now logged out in all top level functions when
an error occurs. The line length and indentation have been improved
and a few variables have been renamed to shorter lines.

Regarding a possible contribution to the Zephyr RTOS project, all
typedef constructs for structured data types have been removed. This
design pattern is not welcomed in source code for Zephyr and could
already lead to a formal rejection.

Signed-off-by: Stephan Linz <linz@li-pro.net>