drivers: mfd_sc16is75x: refactoring MFD core
From the init-function, duplicate code for the configuration
of GPIO pins was converted into a universal auxiliary function.
The same procedure was applied to the hardware reset sequence
and the interrupt binding. The result code is now logged out in
all top level functions when an error occurs.

The line length and indentation have been improved.

Signed-off-by: Stephan Linz <linz@li-pro.net>