drivers: Create driver for DRV84xx series stepper motor drivers
This driver uses a PWM output as its source of stepping pulses. As this
prevents the accurate step counting that would be needed for positioning
actions, only velocity and acceleration modes are supported. Very low stepping
rates may also be impossible depending on the limitations of the underlying PWM
source.