[WIP] lib: Create motion control library
This library abstracts over the underlying stepper action stream subsystem and
provides simple, non-blocking calls to queue basic motion tasks, expressed in
terms of effective physical units (e.g. degrees and seconds instead of steps
and microseconds).

While it is currently built to interact with the stepper action subsystem, it
could be extended to work with any number of similar streaming interfaces.