doc: bridle: revert monkey patch for doxygengroup
Because of a really stupid and shirt-sleeved monkey patch of the
original `doxygengroup` directive (no matter if from the Sphinx
extension `breathe` or `zephyr.doxybridge`) within the Sphinx
extension `zephyr.domain`, the Bridle documentation has to get
back the original behavior of Breathe by overwrite again the
directive entry point with the original class from Breathe for
its own. This is an absolute stopgap solution until the Bridle
documentation has completely abandoned the use of the now obsolete
and no longer maintained "breathe" project and switched to the
Zephyr upstream concept.

Point of origin in Zephyr upstream source code:

- zephyrproject-rtos/zephyr@c8c3a3b: from now on, the monkey patch
  replaces class `DoxygenGroupDirective` with the reimplementation
  from the `zephyr.doxybridge` extensions and not anymore from the
  supposed one from Breathe.

Unfortunately, the Sphinx extension `zephyr.domain` cannot be
omitted because Intersphinx cannot interpret the domain-specific
references (roles). It may even be that this is a design defect
within Sphinx itself.

Signed-off-by: Stephan Linz <linz@li-pro.net>