doc: fix version string handling
Extends the `version` variable with the extra string so that the
HTML builder together with the corresponding theme can output the
correct version number. The new variable `shortversion`, also
integrated as a Sphinx replacement, can now be used within the
Bridle documentation for the correct output from the
`bridle version` command.

Signed-off-by: Stephan Linz <linz@li-pro.net>