boards: bluepill_f103cb: use external crystals
Enables external 8MHz crystal on HSE and use the PLL for upscaling
to 72MHz SYSCLK. Set AHB, APB1 and APB2 prescaler for 72MHz on
Cortex-M system timer (systick), HCLK (DMA) and FCLK and also on
both peripheral buses APB1 and APB2. Set ADC prescaler for 12MHz
on ADC. Set USB prescaler for 48MHz on USB.

Enables external 32768kHz crystal on LSE and enable the on-chip
RTC. Tested with 'timer' shell command over the counter API.

Enables the "Independent Watchdog" (IWDG), but does not use it.

Signed-off-by: Stephan Linz <linz@li-pro.net>