tests: exclude more platforms with sysbuild enabled
Some platforms already have the sysbuild flag set and as long as our
tests are not ready or prepared for the new sysbuild feature, these
platforms need to be explicitly filtered out by our quarantine file.

Unfortunately, there is no way to filter out these platforms with any
of the Twister filter options in the testcase.yaml files.

Signed-off-by: Stephan Linz <linz@li-pro.net>