boards: picoboy: support the new color plus board
The "PicoBoy Color Plus" is a new extended and in some points optimized
version of the "PicoBoy Color". Based on the new RP2350A SoC it brings
back the on-board acceleration sensor from the original "PicoBoy", add
a single NeoPixel digital RGB-LED and beside the solder pads for add-on
components as on the "PicoBoy Color" ther is now an additional Qwicc /
STEMMA QT compatible connector.

Signed-off-by: Stephan Linz <linz@li-pro.net>