drivers: sc18is604: Implement SPI transfers without inserted delays
On actual hardware we need custom handling around the SPI API to insert
the long delays between bytes that the IC expects. For compatibility
with the emulation framework, however, we would like to perform
transactions as single SPI transceives. For this purpose, provide an
alternative, much simpler, implementation for delay-less transactions.

Signed-off-by: Sarah Renkhoff <sarah.renkhoff@navimatix.de>