tests: drivers: watchdog: remove support of Nucleo WBA52CG board
Nucleo WBA52CG board (nucleo_wba52cg) is not supported anymore since
it is NRND (Not Recommended for New Design) and it is not supported
anymore in the STM32CubeWBA from version 1.1.0 (July 2023). The
migration to Nucleo WBA55CG (nucleo_wba55cg) is recommended and
it could be done without any change.

Signed-off-by: Stephan Linz <linz@li-pro.net>