ci: check compliance: repair Zephyr module data fetch
For inexplicable reasons, check_compliance.py once again accesses
the metadata of Zephyr module YAML files directly by the simplest
means instead of using the tools provided for this purpose from
zephyr/scripts/zephyr_module.py or at least implementing missing
methods there and not somewhere distributed in scripts with a
completely different meaning.

However, the existing parser was also implemented completely
incorrectly and made a completely absurd assumption that was not
documented or specified anywhere. Apart from the fact that the
clearly specified YAML structure is not even remotely considered,
it was completely ignored that key-value pairs in a YAML file can
also have leading spaces, that the inclusion in quotation marks
is an option and cannot be demanded and that the syntax does not
give any specifications for upper or lower case.

Signed-off-by: Stephan Linz <linz@li-pro.net>