loopback_test: tiac_magpie: reduce fast transfer clock
Reduce fast transfer clock to max. 15 MHz, because of missing DMA setup.

As long as the board has no DMA acceleration enabled for SPI transfers
the original intended clock of 50 MHz will be to fast for the new Zephyr
upstream timing testcase. This would be lead to a "Very high latency"
failure, the transfer takes way longer than it should.

Signed-off-by: Stephan Linz <linz@li-pro.net>