drivers: mfd: sc18is604: emul: fix VLA compile warning
An array cannot have zero size, thus the additional new
__ASSERT(rx_len != 0, ...) was add.

ISO 9899:2011 6.7.6.2 (https://www.iso-9899.info/n1570.html#6.7.6):

  > 1 Constraints: If the expression is a constant expression, it shall
  >  have a value greater than zero.

  > 5 If the size is an expression that is not an integer constant
  >   expression: if it occurs in a declaration at function prototype
  >   scope, it is treated as if it were replaced by *; otherwise, each
  >   time it is evaluated it shall have a value greater than zero. The
  >   size of each instance of a variable length array type does not
  >   change during its lifetime.

The above text is true both for a plain array (paragraph 1). For a variable
length array (VLA), the behavior is undefined if the expression's value is
less than or equal to zero (paragraph 5). This is normative text in the
C standard. A compiler is not allowed to implement it differently.

Warning was (treats as error in tests):

  .../bridle/drivers/mfd/sc18is604/mfd_sc18is604_emul.c:
    In function ‘mfd_sc18is604_emul_io_spi’:
  .../bridle/drivers/mfd/sc18is604/mfd_sc18is604_emul.c:263:10:
    warning: unbounded use of variable-length array [-Wvla-larger-than=]
    263 |  uint8_t rx[rx_len];
        |          ^~

Signed-off-by: Stephan Linz <linz@li-pro.net>