boards: mini_usb_rp2350: add 16MB QSPI Flash support
Introduces the two hardware revisions 4MB (default) and 16 MB.
This allows use of equally easily available assemblies with more
Flash memory.

For faster identification of firmware images in the target, this
information is output as an informative logging output when the
logging subsystem is enabled: board configuration used at time
if firmware compilation and the resulting QSPI-Flash memory size.

Signed-off-by: Stephan Linz <linz@li-pro.net>