doc: integration for optional WCH-Link tools
The WCH-Link module can be used for on-target debugging and downloading
of firmware binaries build for QingKe RISC-V (Highland Barley) cores.
For working with the WCH-Link, either minichlink or OpenOCD from MSR
(MounRiver Studio), equipped with a special driver, can be used. Both
tools are natively supported by the Zephyr build system, but have to
install from source code.

This adds a detailed deskriptio to it.

Signed-off-by: Stephan Linz <linz@li-pro.net>