tests: add USB Console Test Bed
Similar to the existent test suites for drivers and shields Bridle
can now protect the growing variations of the USB Console snippet
reflected to the well known and supported boards.

This is an initial commit with most of all known test fixtures but
needs some more efforts in validation against all usable boards. Thus
the Twister test specification comes with a white list of currently
supported platforms. Later unresolvable issues should be maintain by
the quarantine list as already descript in the comments.

Signed-off-by: Stephan Linz <linz@li-pro.net>