ci: github: qa-integration: tolerate warnings
For the time being - AND ONLY FOR THE TIME BEING - all warnings inside of a
Twister run will not lead to errors as we need a development window (time)
until release v4.3.0 to move with the Bridle code base from the legacy USB
stack to the next new one.

Tracking [BUG] issue: #358.

Signed-off-by: Stephan Linz <linz@li-pro.net>