samples: helloshell: test: rework features on MIMXRT1170-EVK
The Cortex-M4 core has limited access to Flash (128kB), thus the
following (mostly useless) features are now disabled:

- EEPROM
- REGULATOR
- RTC
- SENSOR

The FLASH API and Shell command is now back and also usable with
the CM4 core. On Rev. A the on-board 6-axis e-compass (FXOS8700CQ)
needs to disable as long the related SENSOR API is disabled and
even this chip was never present on-board since Rev. B.

Signed-off-by: Stephan Linz <linz@li-pro.net>