manifest: support WPA supplicant and Host AP driver
WPA supplicant was designed to be portable for different drivers
and operating systems. It is an implementation of the WPA Supplicant
component, i.e., the part that runs in the client stations. It
implements WPA key negotiation with a WPA Authenticator and EAP
authentication with Authentication Server. In addition, it controls
the roaming and IEEE 802.11 authentication/association of the wlan
driver: WPA-PSK, WPA with EAP, RSN/WPA2, key management for CCMP,
TKIP, WEP104, WEP40. Originally, hostapd was an optional user space
component for Host AP driver.

Signed-off-by: Stephan Linz <linz@li-pro.net>