ci: github: doc: free disk space on start
For a few days now, building the documentation has been reaching
the natural limits of the disk size provided by GitHub for this
workflow. However, since the build depends exclusively on Ubuntu
system packages and the Zephyr SDK stored in the cache, it is
possible to completely delete the tools provided by GitHub for
various scripting languages, .Net, or Android, thereby
significantly increasing disk capacity.

Signed-off-by: Stephan Linz <linz@li-pro.net>