ci: github: doc: go back to run on ubuntu-22.04
With the growing number of boards, the runtime for creating the
board catalog in the final documentation increases. For reasons
that are currently unexplained, this directly leads to unusual
workflow interruptions on GitHub's server when the workflow is
executed on their Linux host. Error message:

    "The operation was canceled." or "The hosted runner lost
    communication with the server. Anything in your workflow
    that terminates the runner process, starves it for
    CPU/Memory, or blocks its network access can cause
    this error."

At the moment, the phenomenological solution is to use a Linux
host based on the older Ubuntu LTS 22.04, which may have something
to do with the lower usage of this type of host. It remains to be
seen what the right solution will be in the long term. The move
to a self-hosted runner should also be considered.

Signed-off-by: Stephan Linz <linz@li-pro.net>